requirements.txt
Flask==2.3.1
Flask_Login==0.6.2
Flask_MySQLdb==1.0.1
flask_sqlalchemy==3.0.3
Flask_WTF==1.1.1
python-dotenv==1.0.0
Werkzeug==2.3.1
WTForms==3.0.1
# manual additions.
PyMySQL==1.0.3
mysql-connector
email_validator==2.0.0
feedgen==0.9.0
requests==2.25.1
beautifulsoup4==4.8.2
pytz==2022.1
reportlab==4.0.4
wordcloud==1.9.2